+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -137,7 +137,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-24T14:42:39.367Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-24T15:35:00.800Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")