+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVX1B4JDH6894TARTJ0FYDJ5"
+        "01KVX4B6A46RBH20XYX2FKD5SZ"
     )
     r"""The unique identifier of the data source"""