+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -2057,7 +2057,7 @@ class AgentToolInputRunTools(BaseModel):
         StreamRunAgentAgentToolInputRunAgentsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KVX1B45D11116VZDHAZVT9F6"
+    id: Optional[str] = "01KVX4B5V2F7CEZBA00EWGQMCV"
 
     description: Optional[str] = None