+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVX1B4JH69Y9CF064RHMTP66"
+        "01KVX4B6AB88PC9RJ51T2EC462"
     )
     r"""The unique identifier of the data source"""